# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Langfuse Light API
  description: |
    Lightweight open-source LLM observability platform — API reference.
    Compatible with Langfuse Python/JS SDKs via the `/api/public` endpoints.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /api/organizations:
    post:
      summary: Create organization
      tags: [Organizations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Organization created
    get:
      summary: List user's organizations
      tags: [Organizations]
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
  /api/projects:
    post:
      summary: Create project
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Project created
    get:
      summary: List user's projects
      tags: [Projects]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
  /api/traces:
    post:
      summary: Create or update a trace
      tags: [Traces]
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTraceBody'
      responses:
        '201':
          description: Trace created
    get:
      summary: List traces
      tags: [Traces]
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: name
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated trace list
  /api/traces/{id}:
    get:
      summary: Get trace with observations
      tags: [Traces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trace detail with observations
  /api/observations:
    post:
      summary: Create an observation
      tags: [Observations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateObservationBody'
      responses:
        '201':
          description: Observation created
...
  /api/ingestion:
    post:
      summary: Batch ingestion (traces + observations)
      tags: [Traces]
      parameters:
        - name: project_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                traces:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateTraceBody'
                observations:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateObservationBody'
      responses:
        '201':
          description: Ingestion result
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    APIKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        created_at:
          type: string
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        org_id:
          type: string
        created_at:
          type: string
    CreateTraceBody:
      type: object
      required: [id]
      properties:
        id:
          type: string
        name:
          type: string
        input:
          type: object
        output:
          type: object
        metadata:
          type: object
        user_id:
          type: string
        session_id:
          type: string
        tags:
          type: array
          items:
            type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        total_cost:
          type: number
        token_usage:
          type: object
...
    CreateObservationBody:
      type: object
      required: [trace_id]
      properties:
        id:
          type: string
        trace_id:
          type: string
        type:
          type: string
          enum: [SPAN, GENERATION, EVENT]
        name:
          type: string
        input:
          type: object
        output:
          type: object
        metadata:
          type: object
        model:
          type: string
        model_parameters:
          type: object
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        token_usage:
          type: object
        cost:
          type: number
        status:
          type: string
        parent_observation_id:
          type: string
...